solution/requirements.txt
# Core Dependencies for Hybrid RAG System
# All versions pinned to latest stable releases (verified Dec 31, 2025)

# Search Engine
# Typesense Server v29.0 compatible with typesense-python >= 0.20.0
# Latest: 1.3.0 (released Nov 14, 2025)
typesense==1.3.0

# Data Processing
# Polars: Fast dataframe library for CSV operations
# Latest: 1.32.3 (stable)
polars==1.33.1

# Markdown parsing for advisory documents
markdown==3.7

# HTML/XML parsing
beautifulsoup4==4.12.3

# LLM & Embeddings
# Google Gemini API client
# Latest: 1.35.0+ (latest)
google-genai>=1.35.0

# Sentence Transformers: State-of-the-art embeddings
# Latest: 3.1.1 (released Nov 2025, requires transformers>=4.41.0)
sentence-transformers==3.1.1

# PyTorch: Deep learning framework (required by sentence-transformers)
# Latest: 2.5.1 (stable, Dec 2025)
# Install CPU-only: pip install torch --index-url https://download.pytorch.org/whl/cpu
torch==2.5.1

# NumPy: Numerical computing
# Latest: 2.1.3 (Dec 2025, requires Python >=3.9)
numpy==2.1.3

# Transformers: NLP models (required by sentence-transformers)
# Latest: 4.46.0+ (required >= 4.41.0)
transformers>=4.46.0

# Utilities
# Environment variable management
python-dotenv==1.0.1

# Terminal output formatting
# Rich: Beautiful terminal output with markdown support
rich==13.8.0
